move table dump and metadata to their own files
